chore: production-ready audit + broader test fixtures
Audit fixes:
- README and docs now show actual usage (quickstart, supported inputs,
  agreement-with-pyiqa note); previously empty of content.
- _resize.py: drop dead scale>=1 branch (only ever called with scale=0.5);
  hard-code the 0.5 specialisation, widen dtype annotation.
- _gaussian.py: privatise KERNEL_SIZE / SIGMA module attributes.
- _features.py: extract log_eta intermediate for readability.
- __init__.py: drop redundant np.float32(255.0) cast.
- Drop `from __future__ import annotations` everywhere (Python 3.14).
- Move PLAN.md from repo root to docs/internal/.
- Add negative-shape test for _to_luma's ValueError path.

Test fixtures:
- New tools/build_test_fixtures.py downloads 3 fresh ~2.7 MP Picsum
  sources and produces 4 quality-degraded variants of each:
  jpeg q10, webp q10, 4x blocky upscale (NEAREST), 4x soft upscale
  (LANCZOS). 15 new images (~4 MB), kept alongside the 5 existing
  natural-image baselines.
- Test glob widened to .jpg / .png / .webp.
- BRISQUE score range across the test set now spans 9-87 (close-to-
  pristine through severely degraded). All 25 cases agree with pyiqa
  within 1.0 BRISQUE points (median 0.04).

100% line coverage; ruff, ruff-format, ty all clean.

+++ pyteenybrisque/__init__.py
@@ -1,7 +1,5 @@
 """Minimal numpy-only BRISQUE no-reference image quality scorer."""
 
-from __future__ import annotations
-
 from os import PathLike
 
 import numpy as np
@@ -34,9 +32,7 @@ def _to_luma(image: object) -> npt.NDArray[np.float32]:
         raise ValueError(f"unsupported image shape {arr.shape}")
     arr = arr[..., :3]
 
-    rgb01 = (
-        arr.astype(np.float32) / np.float32(255.0) if np.issubdtype(arr.dtype, np.integer) else arr.astype(np.float32)
-    )
+    rgb01 = arr.astype(np.float32) / 255.0 if np.issubdtype(arr.dtype, np.integer) else arr.astype(np.float32)
     luma01 = rgb01 @ _LUMA_RGB
     return np.round(luma01 * 255.0).astype(np.float32)
